Extract shared layout into `ItemPage` component
